Refactor(bbs): domain 및 dto 패키지 분리
