+++ src/main/java/egovframework/let/cop/bbs/domain/repository/BBSLoneMasterDAO.java
@@ -1,4 +1,4 @@
-package egovframework.let.cop.bbs.repository;
+package egovframework.let.cop.bbs.domain.repository;
 import java.util.List;
 
 import org.egovframe.rte.psl.dataaccess.EgovAbstractMapper;
